refactor(websites): split componentes monolíticos del frontend
- sections.tsx (1679 líneas) → 12 archivos en sections/ con barrel export
  - HeroSection, AboutSection, ServicesSection, ProductsSection,
    TestimonialsSection, PricingSection, GallerySection, FAQSection,
    ContactSection, GenericSection + shared utilities + BentoGrid
- LivePreview.tsx (1218 líneas) → 487 líneas + styleOverrides.ts (512 líneas)
  - Extraídas 8 funciones de inyección de estilos al iframe
- Backward compatible: sections.tsx re-exporta desde sections/index

Closes #244 (batch 1)
